added all files
this is a full website where you can embed you privet youtube video with youtube api3. it will show all the video you added to this site in i single page (index.php) with pagination . when you click a video you will redirect to video page. and it will ask for a pass to play this video.